backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.10.2
httpx==0.28.1
python-dotenv==1.0.1
crawl4ai
beautifulsoup4
playwright


